requirements.txt
Scrapy
psycopg2-binary
scrapyd
setuptools
scrapyd-client
selenium
webdriver-manager
python-decouple
